feat: add new processActionLinks util (#1331)
adds new processActionLink and processActionLinks utils to "pre-process" a single HubActionLink or an array of HubActionLinks. For now, this involves transforming action links of kind = "content" into action links with an href that can be easily consumed in the UI. We can theoretically add additional "pre-processing" transforms to this util as they become necessary.