feat(hub-common): revamp getDownloadFlow() and canUseCreateReplica() helpers to account for enterpri
affects: @esri/hub-common

To start unlocking the ability to use createReplica for enterprise site downloads, we needed to make
a lot of high
level changes. For example, we introduced the “portalCreateReplica” download flow as
well as the use of a “download
flow availability hash” that is used to calculate which download
flows can be used in the current context. However,
many more changes were required to make this
happen

Below are a list of important changes in this commit:
- The “createReplica" download flow
type was removed for being too ambiguous
- The new “hubCreateReplica” download flow type indicates
that createReplica will be hit via the Hub Download API
- The new “portalCreateReplica” download
flow type indicates that createReplica will be hit directly (i.e., no Hub Download API
intermediary)
- Created the getAvailableDownloadFlows() helper. The returned hash indicates which
download flows are available in the current environment
- getDownloadFlow() accepts a
download-flow-availability hash that it uses to calculate the resulting download flow
- Download
configuration logic now treats configurations for the “hubCreateReplica” and “portalCreateReplica”
download flows interchangeably
- Download configurations that have the "createReplica" download flow
saved will be migrated to “hubCreateReplica”
- canUseCreateReplica() will now return true for main
feature service items in enterprise contexts

BREAKING CHANGE:
- getDownloadConfiguration() now requires an “availableDownloadFlows” argument. Use the output of
the new getAvailableDownloadFlows() helper